requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
execnet==2.1.1
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
playwright==1.48.0
pluggy==1.5.0
pyee==12.0.0
pytest==8.3.3
pytest-base-url==2.1.0
pytest-env==1.1.5
pytest-html==4.1.1
pytest-json-report==1.5.0
pytest-metadata==3.1.1
pytest-playwright==0.5.2
pytest-xdist==3.6.1
python-dotenv==1.0.1
python-slugify==8.0.4
requests==2.32.3
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==2.2.3
